ft_printf

+++ ft_printf.h
@@ -1,3 +1,15 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   ft_printf.h                                        :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: zmourid <marvin@42.fr>                     +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2023/11/25 13:56:49 by zmourid           #+#    #+#             */
+/*   Updated: 2023/11/25 13:56:51 by zmourid          ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
 #ifndef FT_PRINTF_H
 # define FT_PRINTF_H
 # include <stdarg.h>
@@ -7,11 +19,11 @@
 # include <unistd.h>
 
 int	ft_printf(const char *format, ...);
-size_t	ft_strlen(const char *s);
+int	ft_strlen(const char *s);
 int	ft_putnbr(int n);
 int	ft_putstr(char *s);
 int	ft_putchar(int c);
 int	ft_puthex(unsigned int nb, char c);
 int	ft_putaddress(unsigned long n);
-int	ft_putnbr_unsigned(unsigned int  n);
+int	ft_putnbr_unsigned(unsigned int n);
 #endif
